Remove discovery management.
The discovery manager is stored in the user session to keep track
of the current discovered service when a user url yields several
of them. This allows calling Consumer.begin several times to do
a separate authentication process for each service. It is never
used in practice and is better implemented outside the library.